// src/lib/github.types.ts
export type OctokitResponse<T> = {
	data: T;
	status?: number;
	headers?: Record<string, string>;
};

export enum RepoNames {
	image = "code-gen-image-sync",
	sync = "code-gen-sync",
}